break out cilogon ldap stuff into separate module
inspired by SOFTWARE-4528, which uses cilogon info for purposes
unrelated to oasis managers

Also, move safe_dict_get into common.  (Currently it's only used in
oasis_managers, but it is a generic helper function with perhaps
broader applicability.)
